+++ app/src/main/java/woowacourse/paint/model/Brushes.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.customVeiw
+package woowacourse.paint.model
 
 import android.graphics.Canvas
