+++ app/src/main/java/woowacourse/paint/ui/ColorAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.ui
 
 import android.graphics.Color
 import android.view.ViewGroup
