+++ app/src/main/java/woowacourse/paint/ui/ColorViewHolder.kt
@@ -1,10 +1,11 @@
-package woowacourse.paint
+package woowacourse.paint.ui
 
 import android.graphics.Color
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import woowacourse.paint.R
 
 class ColorViewHolder(itemView: View, onClickListener: (Int) -> Unit) :
     RecyclerView.ViewHolder(itemView) {
